Add files via upload
The Car Showroom Management System is a Java-based application designed to manage details related to car showrooms. It allows users to add and retrieve information such as showroom details, employee records, and car details. The system offers basic functionalities to input new data and view existing entries, streamlining the management of essential information for a car showroom. This project serves as a simple yet efficient solution for organizing and accessing car showroom data.